Fixed: 영속성 이슈 처리 - Entity -> DTO 변환 후, 영속화가 풀리게 되면서 Save 메서드를 호출할 때, JPA Exception이 발생 - infrastructure 계층에서 Entity만 사용하도록 변경
