+++ src/main/kotlin/com/hhplus/enrollment/lecture/domain/LectureServiceImpl.kt
@@ -5,8 +5,10 @@ import com.hhplus.enrollment.lecture.domain.validator.LectureValidator
 import com.hhplus.enrollment.lecture.infrastructure.LectureHistoryRepository
 import com.hhplus.enrollment.lecture.infrastructure.LectureRepository
 import com.hhplus.enrollment.lecture.infrastructure.dto.LectureHistoryCommandDto
-import jakarta.transaction.Transactional
+import com.hhplus.enrollment.lecture.infrastructure.dto.LectureQueryDto
 import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Propagation
+import org.springframework.transaction.annotation.Transactional
 
 @Service
 class LectureServiceImpl(
@@ -24,19 +26,19 @@ class LectureServiceImpl(
         return result.map { LectureData.from(it) }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     override fun enroll(query: LectureCommandData): LectureData {
         validator.validate(query) // 수강자 정보, 신청 이력에 대한 유효성 검사
 
-        val lecture = getLecture(LectureQueryData.of(query.lectureId)) // 강의 정보 조회. 내부에서 유효성 검사
+        val lecture = lectureRepository.getLecture(LectureQueryDto(query.lectureId)) // 강의 정보 조회. 내부에서 유효성 검사
         lecture.enroll() // 수강 신청. 내부에서 유효성 검사
 
-        lectureRepository.enroll(lecture.toDto()) // 참감 된 수강 정보 저장
+        lectureRepository.enroll(lecture) // 참감 된 수강 정보 저장
 
         // 강의 이력 저장
         val historyCommand = LectureHistoryCommandDto.of(query.lectureId, query.traineeId)
         lectureHistoryRepository.insert(historyCommand)
-        return lecture
+        return LectureData.from(lecture)
     }
 
     override fun getLectureHistories(query: LectureHistoryQueryData): List<LectureHistoryData> {
