+++ src/main/kotlin/com/hhplus/enrollment/lecture/infrastructure/LectureRepository.kt
@@ -1,12 +1,12 @@
 package com.hhplus.enrollment.lecture.infrastructure
 
-import com.hhplus.enrollment.lecture.infrastructure.dto.LectureDto
 import com.hhplus.enrollment.lecture.infrastructure.dto.LectureQueryDto
+import com.hhplus.enrollment.lecture.infrastructure.jpa.entity.LectureEntity
 
 interface LectureRepository {
-    fun getLecture(query: LectureQueryDto): LectureDto
-    fun getAvailableLectures(): List<LectureDto>
+    fun getLecture(query: LectureQueryDto): LectureEntity
+    fun getAvailableLectures(): List<LectureEntity>
 
-    fun enroll(command: LectureDto): LectureDto
+    fun enroll(command: LectureEntity): LectureEntity
 
 }
